Add WAD signatures for PAL, use temp file for subfiles

+++ Spyro Editor/Models/WADModel.cs
@@ -59,7 +59,7 @@ namespace Spyro_Editor.Models
             return WAD!;
         }
 
-        public Subfile? GetSubfile(byte id)
+        public Subfile? GetSubfile(short id)
         {
             return WAD!.Subfiles.Find(s => s.Id == id);
         }
